Relocate FileCommands into headless Gum.Presentation, completing #3862
Extracted IFileWatchIgnoreList out of its co-located FileWatchIgnoreList.cs (was declared inline with the concrete class, both still in Gum.csproj) into its own file in Gum.Presentation - the concrete class stays in Gum.csproj.

Added a narrow IRecycleBinService seam (same shape as IGumCursorState/IResizeHandlesVisual) for MoveToRecycleBin - the only genuinely stuck piece, since Microsoft.VisualBasic.FileIO's recycle-bin API needs the Windows desktop runtime (UseWindowsForms). Concrete RecycleBinService stays in Gum.csproj.

Added an ICsvLocalizationLoader seam for the CSV-loading branch of LoadLocalizationFile - AddDatabaseFromCsv depends on CsvLibrary, which targets net8.0-windows and can't be referenced from the net8.0 Gum.Presentation assembly (a hard NuGet TFM-compatibility error, not just an WPF/WinForms flag). Concrete CsvLocalizationLoader (a thin wrapper around the existing extension method) stays in Gum.csproj.

Inlined two Locator-based/concrete-class calls using dependencies FileCommands already has injected: GetFullPathXmlFile's ElementSave extension method (mirrors an identical fix already done in DragDropManager) now uses the injected IProjectManager instead of Locator.GetRequiredService; ProjectManager.IsFileReadOnly (a pure, stateless one-liner) is now a local private duplicate instead of a call into the concrete, still-tool-side ProjectManager class.

Tests: split FileCommandsTests - the one test needing real CSV parsing (CsvLibrary can't be referenced from headless test project either) stays in GumToolUnitTests wired to the real CsvLocalizationLoader; everything else (11 existing + 2 new: MoveToRecycleBin delegation, GetFullFileName path computation) moved to Gum.Presentation.Tests. Both new tests mutation-tested (breaking each call/logic turned its test red; reverted).

This completes all four classes named in #3862's original scope (ElementCommands #3867, EditCommands/ProjectCommands #3869, FileCommands here).
